refactor: move middleware & types to the lib directory
these should not exist on the same level as domains (provider,
support-resource)
